use API_URL as var
